Resolve various integer conversion issues
While only warnings with GCC, those were actually fatal with Clang.

+++ src/lib/rule_p.h
@@ -136,9 +136,9 @@ public:
 class Date
 {
 public:
-    uint16_t year;
-    uint8_t month;
-    uint8_t day;
+    int year;
+    int month;
+    int day;
     enum VariableDate : uint8_t {
         FixedDate,
         Easter
